// src/app/shared/model/sotheodoi.model.ts
export class SotheodoiModel {
    id: number;
    sohoso: string;
    patient_id: number;
    mabenhnhan: string;
    sotheodoi_js: string;
}
// create table `sotheodoi` (
// 	`id` bigint(20) NOT NULL AUTO_INCREMENT,
//     `sohoso` varchar(15),
//     `patient_id` bigint(20) not null,
//     `mabenhnhan` varchar(15),
//     `sotheodoi_js` text(50000),
//     PRIMARY KEY (`id`)
//   )
// engine=InnoDB auto_increment=4 default charset=utf8;